video recoding

+++ webcam-server.rb
@@ -1,6 +1,5 @@
 require 'fileutils'
 def cleanup 
-	Process.kill("HUP", @@ngrok_pid) unless @ngrok_pid.nil?
 	FileUtils.rm_r "segments"
 	system("killall ffmpeg")
 end
@@ -18,9 +17,10 @@ require "video.rb"
 
 @@config = {}
 @@tokens = {}
-@@ngrok_pid = nil
 
 @@segment_ctr = 0
+
+@@mode = :picture
 
 
 def reload_config
@@ -90,13 +90,10 @@ def record_local(token, res)
 end
 
 def main_url
-	p @@config
-	"/#{@@config['main_token']}.html"
-end
-
-def setup_ngrok
-	if @@config['expose_via_ngrok']
-		@@ngrok_pid = spawn("ngrok http #{@@config['port']}")
+	if @@mode == :picture
+		"/#{@@config['main_token']}.html"
+	elsif @@mode == :video
+		"/#{@@config['main_token']}.m3u8"
 	end
 end
 
@@ -115,8 +112,11 @@ if ARGV.size > 0
 	if arg == "list"
 		puts "List of devices\n"
 		puts list_webcams
+		exit
 	end
-	exit
+	if arg == "video"
+		@@mode = :video
+	end
 end
 
 if video_streaming_enabled?
@@ -129,7 +129,6 @@ end
 
 set :port, @@config['port']
 set :bind, "0.0.0.0"
-setup_ngrok
 
 
 configure do
@@ -149,36 +148,68 @@ get '/reload/:reload_token' do
 	redirect(main_url)
 end
 
-get '/:token.html' do
-	return status 403 unless token_valid?(params[:token], {ip: request.ip})
-	@@html_code.to_s.gsub("<TOKEN>", params[:token])
-end
+if @@mode == :picture 
+	get '/:token.html' do
+		return status 403 unless token_valid?(params[:token], {ip: request.ip})
+		@@html_code.to_s.gsub("<TOKEN>", params[:token])
+	end
 
-get '/:token/image.jpg' do
-	return status 403 unless token_valid?(params[:token], {ip: request.ip})
-	filename = filename_for(params[:token])
-	w = params[:w] || width
-	h = params[:h] || height
-	capture(w, h, webcam_name, filename)
-	res = File.read(filename).to_s
-	record_local(params[:token], res)
-	File.delete(filename)
-	res
+	get '/:token/image.jpg' do
+		return status 403 unless token_valid?(params[:token], {ip: request.ip})
+		filename = filename_for(params[:token])
+		w = params[:w] || width
+		h = params[:h] || height
+		capture(w, h, webcam_name, filename)
+		res = File.read(filename).to_s
+		record_local(params[:token], res)
+		File.delete(filename)
+		res
+	end
 end
 
 # Video URLS
-get '/:token.m3u8' do
-	return status 404 unless video_streaming_enabled?
-	return status 503 unless video_available?
-	return status 403 unless token_valid?(params[:token], {ip: request.ip})
-	video_playlist(@@segment_ctr, params[:token], (params[:size].to_i || PLAYLIST_LENGTH))
-end
+if @@mode == :video
+	get '/:token.m3u8' do
+		return status 404 unless video_streaming_enabled?
+		return status 503 unless video_available?
+		return status 403 unless token_valid?(params[:token], {ip: request.ip})
+		video_playlist(@@segment_ctr, params[:token], params[:size].to_i)
+	end
 
-get '/:token/video_:segment.mpg' do
-	return status 404 unless video_streaming_enabled?
-	return status 403 unless token_valid?(params[:token], {ip: request.ip})
-	res = File.read(segment_filename(params[:segment])).to_s
-	res
-end
+	get '/:token/video_:segment.mpg' do
+		return status 404 unless video_streaming_enabled?
+		return status 403 unless token_valid?(params[:token], {ip: request.ip})
+		res = File.read(segment_filename(params[:segment])).to_s
+		res
+	end
 
+	get '/:token/:sfrom/:sto/video.mpg' do
+		return status 404 unless video_streaming_enabled?
+		return status 403 unless token_valid?(params[:token], {ip: request.ip})
+		from = params[:sfrom].to_i
+		to = params[:sto].to_i
+		out = "video_#{from}_#{to}.mpg"
+		success = video_concatenate(from, to, out)
+		
+		return status 404 unless success
+		
+		res = File.read(out).to_s
+		File.delete(out)
+		res
+	end
+
+	get '/:token/last_seconds/:seconds.mpg' do
+		return status 404 unless video_streaming_enabled?
+		return status 403 unless token_valid?(params[:token], {ip: request.ip})
+		seconds = params[:seconds].to_i
+		n_segments = (seconds.to_f / SECONDS_PER_SEGMENT.to_f).to_i
+		from = @@segment_ctr - n_segments
+		to = @@segment_ctr - 1
+		redirect("/#{params[:token]}/#{from}/#{to}/video.mpg")
+	end
+
+	get '/:token/last_minutes/:minutes.mpg' do
+		redirect("/#{params[:token]}/last_seconds/#{params[:minutes].to_i*60}.mpg")
+	end
+end
